backend/requirements.txt
asgiref==3.9.1
click==8.1.8
coverage==7.10.7
dj-database-url==3.0.1
Django==4.2.24
django-cors-headers==4.9.0
djangorestframework==3.16.1
exceptiongroup==1.3.0
gunicorn==23.0.0
h11==0.16.0
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==8.4.2
python-decouple==3.8
sqlparse==0.5.3
tomli==2.3.0
typing_extensions==4.15.0
uvicorn==0.38.0
whitenoise==6.11.0
